utils: Fix class_weight dtype

+++ elektronn3/data/utils.py
@@ -83,7 +83,7 @@ def calculate_class_weights(
             np.sum(np.equal(targets, c))
             for c in classes
         ], dtype=np.float32)
-        class_weights = targets.size / num_labeled
+        class_weights = (targets.size / num_labeled).astype(np.float32)
         return class_weights
 
     def __binmean(targets):
